+++ packages/viness-core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @viness/di
 
+## 1.5.2
+
+### Patch Changes
+
+- refactor: rename UIState to UIStore
+- fix: route path genation error
+
 ## 1.5.1
 
 ### Patch Changes